抽取公共逻辑,改进注册

+++ src/main/java/c4/champions/common/affix/affix/AffixAdaptable.java
@@ -40,7 +40,7 @@ public class AffixAdaptable extends AffixBase {
         DamageType damageType = AffixNBT.getData(cap, this.getIdentifier(), DamageType.class);
 
         if (damageType.name.equalsIgnoreCase(type)) {
-            newAmount -= amount * ConfigHandler.affix.adaptable.damageReductionIncrement * damageType.count;
+            newAmount -= (float) (amount * ConfigHandler.affix.adaptable.damageReductionIncrement * damageType.count);
             damageType.count++;
         } else {
             damageType.name = type;
